Test deploy/teardown back to comments instead of deployments
The deployments system seemed like a bit of a headache in terms of managing all the deployments, making them inactive on teardown, etc.

So the comments system was probably good enough, and hopefully has fewer issues

Also deployments might have been the cause of the test-deploy workflow triggering twice, which was another issue. I am not sure if that is the case, but now we can test that.

(The hypothesis is that the deployment triggered by the PR reopening was triggering a "PR synchronize" event, which in turn was re-triggering the workflow that does the deployment)