Test deploy try setting concurrency at workflow level
It seems like the deployments were NOT the cause of the issue regarding the workflow triggering twice. Without it, re-opening the PR seems still trigger twice.

It looks like it both detects the PR re-opening as one event, and the new commits since the PR was closed being added to the PR as another (synchronize) event, which triggers the workflow twice.

Hopefully, setting the concurrency at a workflow level rather than only in the deploy job makes it so the extra run never happens, instead of being cancelled (which shows as an error)