Test deploy fix get workflow_id from run
It turns out that context.workflow does not return the workflow ID, but rather the workflow name (yes, the name shown in the UI. I don't understand why, but that's the case). And it does not seem to be a workflow_id property available from the context, so we get the workflow run object from the GitHub API, which does contain a workflow_id property.