Fix all relative links to other docs
Explanation: When you make a link to another document, you can either link to a relative path, which is taken literally and not touched by Docusaurus (which leads to errors in this case), or you can link to the markdown file (with the .md extension). Docusaurus will detect markdown file links and convert them to the proper URL path.
Source: https://docusaurus.io/docs/markdown-features/links