Fix broken links (#3)
* Fix footer links - prepend needed docs path

Because those links could go to any part of the website (docs, blog, etc), we need to specify the full subpath, including the "docs/" path first

* Fix footer links 2 - start with slash (/)

Maybe they need to start with a slash?

* Fix footer intro link

If the file is named 'intro.md' in the 'intro' folder, the path should be 'docs/intro', not 'docs/intro/intro', because it seems that Docusaurus is able to detect that the file is named the same as the folder, so it won't double the name in the URL.

* Simplify FAQ path by renaming the doc

* Fix FAQ links

Oops, the updated FAQ links did not commit properly before

* Fix new FAQ docId

docId is "faq/faq", but URL is "/docs/faq". Important to remember the difference between the docId and the URL path

* Add Process Miner link to footer

* Fix all relative links to other docs

Explanation: When you make a link to another document, you can either link to a relative path, which is taken literally and not touched by Docusaurus (which leads to errors in this case), or you can link to the markdown file (with the .md extension). Docusaurus will detect markdown file links and convert them to the proper URL path.
Source: https://docusaurus.io/docs/markdown-features/links

* Remove (seemingly) duplicate Spanish translations

* Rename FAQ document in Spanish translation

* Fix relative links to docs in Spanish translations

* Copied process miner docs to Spanish translation

(TODO: Actually translate process miner docs to Spanish - #4)

* Process Miner knowledge fix broken link

* Fix missing broken link in main-perspective

* Process Miner knowledge fix broken link (english)

In the other commit apparently I only fixed the spanish version

* Fix Process Miner runner-properties wrong links

They were anchors instead of links to files in the same folder for some reason

* Fix anchor Process Miner create-an-installer

At some point, probably the section title was renamed but the anchor links were not updated

* Fix FAQ Spanish version link wrong uppercase

It seems like all warnings are gone - finally!